feat - add accumulator list class for handling lists
