+++ configurations/devices/microphone/microphone_device_config.json
@@ -1,5 +1,6 @@
 {
         "audio_chunk": 1024,
         "rate": 44100,
-        "volume_threshold": 1000
+        "volume_threshold": 1000,
+        "silence_time_threshold": 2
 }
